--- app/src/main/java/com/raheemjnr/fetch_hiring_exercise/navigation/destinations.kt
@@ -1,7 +1,0 @@
-package com.raheemjnr.fetch_hiring_exercise.navigation
-
-
-//sealed class to list out screens that will be available in app
-sealed class MainScreen(val route: String) {
-    object FetchScreen : MainScreen("fetchScreen")
-}
